delete client

+++ actions/deleteClient.ts
@@ -1,13 +1,16 @@
 import { account, config, database } from "@/lib/appwrite"
+import { Query } from "appwrite"
 
 export const deleteClient=async(clientId:string)=>{
   
-
-        const client=await database.deleteDocument(
+       const session=await account.getSession('current')
+      
+       if(!session) throw Error
+       await database.deleteDocument(
             config.databaseId,
             config.clientTable,
             clientId,
+       
         )
-  
-    return client
+       
 }
